+++ src/main/java/net/silentchaos512/mechanisms/api/crafting/recipe/fluid/IFluidRecipeType.java
@@ -1,8 +1,0 @@
-package net.silentchaos512.mechanisms.api.crafting.recipe.fluid;
-
-import net.minecraft.util.ResourceLocation;
-
-@FunctionalInterface
-public interface IFluidRecipeType<T extends IFluidRecipe<?>> {
-    ResourceLocation getTypeName();
-}