+++ src/main/java/net/silentchaos512/mechanisms/block/solidifier/SolidifierScreen.java
@@ -1,0 +1,63 @@
+package net.silentchaos512.mechanisms.block.solidifier;
+
+import net.minecraft.entity.player.PlayerInventory;
+import net.minecraft.util.ResourceLocation;
+import net.minecraft.util.text.ITextComponent;
+import net.silentchaos512.mechanisms.SilentMechanisms;
+import net.silentchaos512.mechanisms.block.AbstractMachineBaseScreen;
+import net.silentchaos512.mechanisms.client.renderer.RenderUtils;
+import net.silentchaos512.mechanisms.util.TextUtil;
+
+public class SolidifierScreen extends AbstractMachineBaseScreen<SolidifierContainer> {
+    public static final ResourceLocation TEXTURE = SilentMechanisms.getId("textures/gui/solidifier.png");
+
+    public SolidifierScreen(SolidifierContainer container, PlayerInventory playerInventory, ITextComponent titleIn) {
+        super(container, playerInventory, titleIn);
+    }
+
+    @Override
+    public ResourceLocation getGuiTexture() {
+        return TEXTURE;
+    }
+
+    @Override
+    public void render(int mouseX, int mouseY, float partialTicks) {
+        this.renderBackground();
+        super.render(mouseX, mouseY, partialTicks);
+        this.renderHoveredToolTip(mouseX, mouseY);
+    }
+
+    @Override
+    protected void renderHoveredToolTip(int mouseX, int mouseY) {
+        if (isPointInRegion(57, 17, 13, 51, mouseX, mouseY)) {
+            renderTooltip(TextUtil.fluidWithMax(container.getFluidInTank(0), SolidifierTileEntity.TANK_CAPACITY).getFormattedText(), mouseX, mouseY);
+        }
+        if (isPointInRegion(153, 17, 13, 51, mouseX, mouseY)) {
+            renderTooltip(TextUtil.energyWithMax(container.getEnergyStored(), container.getMaxEnergyStored()).getFormattedText(), mouseX, mouseY);
+        }
+        super.renderHoveredToolTip(mouseX, mouseY);
+    }
+
+    @Override
+    protected void drawGuiContainerBackgroundLayer(float partialTicks, int mouseX, int mouseY) {
+        super.drawGuiContainerBackgroundLayer(partialTicks, mouseX, mouseY);
+
+        int xPos = (this.width - this.xSize) / 2;
+        int yPos = (this.height - this.ySize) / 2;
+
+        // Progress arrow
+        int progress = container.getProgress();
+        int processTime = container.getProcessTime();
+        int length = processTime > 0 && progress > 0 && progress < processTime ? progress * 24 / processTime : 0;
+        blit(xPos + 79, yPos + 35, 176, 14, length + 1, 16);
+
+        // Energy meter
+        int energyBarHeight = container.getEnergyBarHeight();
+        if (energyBarHeight > 0) {
+            blit(xPos + 154, yPos + 68 - energyBarHeight, 176, 31, 12, energyBarHeight);
+        }
+
+        // Tanks
+        RenderUtils.renderGuiTank(container.getFluidInTank(0), SolidifierTileEntity.TANK_CAPACITY, xPos + 58, yPos + 18, 0, 12, 50);
+    }
+}