+++ src/main/java/net/silentchaos512/mechanisms/block/solidifier/SolidifierTileEntity.java
@@ -1,4 +1,4 @@
-package net.silentchaos512.mechanisms.block.refinery;
+package net.silentchaos512.mechanisms.block.solidifier;
 
 import net.minecraft.entity.player.PlayerInventory;
 import net.minecraft.inventory.container.Container;
@@ -9,7 +9,7 @@ import net.minecraftforge.fluids.FluidStack;
 import net.minecraftforge.fluids.capability.IFluidHandler;
 import net.silentchaos512.mechanisms.AbstractFluidMachineTileEntity;
 import net.silentchaos512.mechanisms.api.IFluidContainer;
-import net.silentchaos512.mechanisms.crafting.recipe.RefiningRecipe;
+import net.silentchaos512.mechanisms.crafting.recipe.SolidifyingRecipe;
 import net.silentchaos512.mechanisms.init.ModTileEntities;
 import net.silentchaos512.mechanisms.util.InventoryUtils;
 import net.silentchaos512.mechanisms.util.MachineTier;
@@ -17,46 +17,15 @@ import net.silentchaos512.mechanisms.util.TextUtil;
 
 import javax.annotation.Nullable;
 import java.util.Collection;
+import java.util.Collections;
 
-public class RefineryTileEntity extends AbstractFluidMachineTileEntity<RefiningRecipe> {
-    public static final int FIELDS_COUNT = 17;
-    public static final int TANK_CAPACITY = 4_000;
-    public static final int ENERGY_PER_TICK = 100;
+public class SolidifierTileEntity extends AbstractFluidMachineTileEntity<SolidifyingRecipe> {
+    public static final int FIELDS_COUNT = 9;
+    public static final int TANK_CAPACITY = 4000;
+    public static final int ENERGY_PER_TICK = 50;
 
-    public RefineryTileEntity() {
-        super(ModTileEntities.refinery, 4, 5, TANK_CAPACITY, MachineTier.STANDARD);
-    }
-
-    @Nullable
-    @Override
-    public RefiningRecipe getRecipe() {
-        if (world == null) return null;
-        return world.getRecipeManager().getRecipe(RefiningRecipe.RECIPE_TYPE, this, world).orElse(null);
-    }
-
-    @Override
-    protected int getProcessTime(RefiningRecipe recipe) {
-        return recipe.getProcessTime();
-    }
-
-    @Override
-    protected Collection<FluidStack> getFluidResults(RefiningRecipe recipe) {
-        return recipe.getFluidResults(this);
-    }
-
-    @Override
-    protected int getEnergyUsedPerTick() {
-        return ENERGY_PER_TICK;
-    }
-
-    @Override
-    protected int getInputTanks() {
-        return 1;
-    }
-
-    @Override
-    protected int getOutputTanks() {
-        return 4;
+    public SolidifierTileEntity() {
+        super(ModTileEntities.solidifier, 3, 1, TANK_CAPACITY, MachineTier.STANDARD);
     }
 
     @Override
@@ -64,7 +33,6 @@ public class RefineryTileEntity extends AbstractFluidMachineTileEntity<RefiningR
         if (world == null || world.isRemote) return;
 
         tryFillTank();
-        tryFillFluidContainer();
 
         super.tick();
     }
@@ -99,49 +67,70 @@ public class RefineryTileEntity extends AbstractFluidMachineTileEntity<RefiningR
                 && (output.isEmpty() || output.getCount() < output.getMaxStackSize());
     }
 
-    private void tryFillFluidContainer() {
-        // Fill empty fluid containers with output fluids
-        ItemStack input = getStackInSlot(2);
-        if (input.isEmpty()) return;
+    @Override
+    protected int getEnergyUsedPerTick() {
+        return ENERGY_PER_TICK;
+    }
 
-        FluidStack fluidInInput = IFluidContainer.getBucketOrContainerFluid(input);
-        if (!fluidInInput.isEmpty()) return;
+    @Override
+    protected int[] getOutputSlots() {
+        return new int[]{2};
+    }
 
-        for (int i = 1; i < 5; ++i) {
-            FluidStack fluidInTank = getFluidInTank(i);
-            if (fluidInTank.getAmount() >= 1000) {
-                ItemStack filled = IFluidContainer.fillBucketOrFluidContainer(input, fluidInTank);
-                if (!filled.isEmpty() && InventoryUtils.mergeItem(this, filled, 3)) {
-                    tanks[i].drain(1000, FluidAction.EXECUTE);
-                    input.shrink(1);
-                    return;
-                }
-            }
-        }
+    @Override
+    protected int getInputTanks() {
+        return 1;
+    }
+
+    @Override
+    protected int getOutputTanks() {
+        return 0;
+    }
+
+    @Nullable
+    @Override
+    public SolidifyingRecipe getRecipe() {
+        if (world == null) return null;
+        return world.getRecipeManager().getRecipe(SolidifyingRecipe.RECIPE_TYPE, this, world).orElse(null);
+    }
+
+    @Override
+    protected int getProcessTime(SolidifyingRecipe recipe) {
+        return recipe.getProcessTime();
+    }
+
+    @Override
+    protected Collection<ItemStack> getProcessResults(SolidifyingRecipe recipe) {
+        return Collections.singletonList(recipe.getCraftingResult(this));
+    }
+
+    @Override
+    protected Collection<FluidStack> getFluidResults(SolidifyingRecipe recipe) {
+        return recipe.getFluidOutputs();
     }
 
     @Override
     public int[] getSlotsForFace(Direction side) {
-        return new int[]{0, 1, 2, 3};
+        return new int[]{0, 1, 2};
     }
 
     @Override
     public boolean canInsertItem(int index, ItemStack stack, @Nullable Direction direction) {
-        return (index == 0 && InventoryUtils.isFilledFluidContainer(stack)) || (index == 2 && InventoryUtils.isEmptyFluidContainer(stack));
+        return index == 0 && InventoryUtils.isFilledFluidContainer(stack);
     }
 
     @Override
     public boolean canExtractItem(int index, ItemStack stack, Direction direction) {
-        return index == 1 || index == 3;
+        return index == 1 || index == 2;
     }
 
     @Override
     protected ITextComponent getDefaultName() {
-        return TextUtil.translate("container", "refinery");
+        return TextUtil.translate("container", "solidifier");
     }
 
     @Override
     protected Container createMenu(int id, PlayerInventory player) {
-        return new RefineryContainer(id, player, this, this.fields);
+        return new SolidifierContainer(id, player, this, this.fields);
     }
 }