+++ src/main/java/net/silentchaos512/mechanisms/compat/jei/SMechanismsJeiPlugin.java
@@ -21,6 +21,7 @@ import net.silentchaos512.mechanisms.block.electricfurnace.ElectricFurnaceContai
 import net.silentchaos512.mechanisms.block.electricfurnace.ElectricFurnaceScreen;
 import net.silentchaos512.mechanisms.block.mixer.MixerScreen;
 import net.silentchaos512.mechanisms.block.refinery.RefineryScreen;
+import net.silentchaos512.mechanisms.block.solidifier.SolidifierScreen;
 import net.silentchaos512.mechanisms.crafting.recipe.*;
 import net.silentchaos512.mechanisms.init.ModBlocks;
 import net.silentchaos512.mechanisms.util.Constants;
@@ -46,6 +47,7 @@ public class SMechanismsJeiPlugin implements IModPlugin {
         registration.addRecipeCategories(new DryingRecipeCategoryJei(guiHelper));
         registration.addRecipeCategories(new MixingRecipeCategory(guiHelper));
         registration.addRecipeCategories(new RefiningRecipeCategory(guiHelper));
+        registration.addRecipeCategories(new SolidifyingRecipeCategory(guiHelper));
     }
 
     @Override
@@ -56,6 +58,7 @@ public class SMechanismsJeiPlugin implements IModPlugin {
         registration.addRecipes(getRecipesOfType(DryingRecipe.RECIPE_TYPE), Constants.DRYING);
         registration.addRecipes(getRecipesOfType(MixingRecipe.RECIPE_TYPE), Constants.MIXING);
         registration.addRecipes(getRecipesOfType(RefiningRecipe.RECIPE_TYPE), Constants.REFINING);
+        registration.addRecipes(getRecipesOfType(SolidifyingRecipe.RECIPE_TYPE), Constants.SOLIDIFYING);
     }
 
     private static List<IRecipe<?>> getRecipesOfType(IRecipeType<?> recipeType) {
@@ -73,6 +76,7 @@ public class SMechanismsJeiPlugin implements IModPlugin {
                 VanillaRecipeCategoryUid.BLASTING, VanillaRecipeCategoryUid.FURNACE);
         registration.addRecipeClickArea(MixerScreen.class, 92, 31, 24, 23, Constants.MIXING);
         registration.addRecipeClickArea(RefineryScreen.class, 43, 31, 24, 23, Constants.REFINING);
+        registration.addRecipeClickArea(SolidifierScreen.class, 79, 31, 24, 23, Constants.SOLIDIFYING);
     }
 
     @Override
@@ -95,5 +99,6 @@ public class SMechanismsJeiPlugin implements IModPlugin {
                 VanillaRecipeCategoryUid.BLASTING, VanillaRecipeCategoryUid.FURNACE);
         registration.addRecipeCatalyst(new ItemStack(ModBlocks.mixer), Constants.MIXING);
         registration.addRecipeCatalyst(new ItemStack(ModBlocks.refinery), Constants.REFINING);
+        registration.addRecipeCatalyst(new ItemStack(ModBlocks.solidifier), Constants.SOLIDIFYING);
     }
 }