+++ CHANGELOG.md
@@ -4,6 +4,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Solidifier, a machine that takes one liquid and outputs items. Recipes can be added with data packs.
+- Recipe to make plastic pellets from polyethylene (solidifier)
+- Recipe to make ice from water (solidifier)
 ### Fixed
 - Might fix another wire-related crash [#23]